requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-redis==4.11.0
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.8
msgpack==0.6.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
